+++ looper/cli_pydantic.py
@@ -246,11 +246,15 @@ def run_looper(args: TopLevelParser, parser: ArgumentParser, test_args=None):
         # Check at the beginning if user wants to use pipestat and pipestat is configurable
         is_pipestat_configured = (
             prj._check_if_pipestat_configured(pipeline_type=PipelineLevel.PROJECT.value)
-            if getattr(args, "project", None)
+            if getattr(subcommand_args, "project", None)
             else prj._check_if_pipestat_configured()
         )
 
         if subcommand_name in ["run", "rerun"]:
+            if getattr(subcommand_args, "project", None):
+                _LOGGER.warning(
+                    "Project flag set but 'run' command was used. Please use 'runp' to run at project-level."
+                )
             rerun = subcommand_name == "rerun"
             run = Runner(prj)
             try: